Module String.LE

Functions reading according to Little Endian byte order

val get_uint16 : t -> int -> int

get_uint16 buff i reads 2 bytes at offset i as an unsigned int of 16 bits. i.e. It returns a value between 0 and 2^16-1

val get_int16 : t -> int -> int

get_int16 buff i reads 2 byte at offset i as a signed int of 16 bits. i.e. It returns a value between -2^15 and 2^15-1

val get_int32 : t -> int -> int32

get_int32 buff i reads 4 bytes at offset i as an int32.

val get_int64 : t -> int -> int64

get_int64 buff i reads 8 bytes at offset i as an int64.