Module 1-Env.Fitness

The fitness of a block is defined as a list of bytes, compared in a lexicographical order (longer list are greater).

include S.T with type t = bytes list

type t = bytes list

include Compare.S with type t := t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val pp : Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val to_bytes : t -> bytes

val of_bytes : bytes -> t option